Invoices are now only viewable by auto transfusionists that have been assigned to them.  This is controlled by the staff information page.  Decimal fields have been updated to allow decimal fields.

+++ resources/views/invoices/create.blade.php
@@ -108,8 +108,8 @@
 		.children("td:nth-child(" + columnIndex + ")")
 		.each(function() 
 		{
-			var value = parseInt($(this).find("input").val());			
-			tot += (isNaN(value) ? 0 : value);
+			var value = parseFloat($(this).find("input").val());			
+			tot += (isNaN(value) ? 0.00 : value);
   		});
 		
   		return tot;
